trackgen: fix visual studio warning
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@9375 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: fdde08be715b878a9eb9ee210a7546ea0f34b635
Former-commit-id: f03681a272186daed0d6086249615260322c2ee6

+++ src/tools/trackgen/ac3d.cpp
@@ -925,7 +925,7 @@ void Ac3d::Object::removeUnusedVertices()
                 for (auto &ref : surface.refs)
                 {
                     if (ref.index == originalIndex)
-                        ref.index = i;
+                        ref.index = static_cast<int>(i);
                 }
             }
             i++;
